Refactor av sidemeny
